refactor: separate functions from get-breadcrumbs.js
