+++ src/utils/get-breadcrumbs.js
@@ -1,12 +1,3 @@
-const getFlatSidebar = (sidebar, path = []) =>
-  sidebar.reduce((acc, item, index) => {
-    const current = { title: item.title, slug: item.slug, path: [...path, index] };
-    if (item.items) {
-      return [...acc, current, ...getFlatSidebar(item.items, current.path)];
-    }
-    return [...acc, { ...item, path: [...path, index] }];
-  }, []);
-
 const getBreadcrumbs = (slug, flatSidebar, sidebar) => {
   const path = flatSidebar.find((item) => item.slug === slug)?.path;
   const arr = [];
@@ -31,5 +22,4 @@ const getBreadcrumbs = (slug, flatSidebar, sidebar) => {
 
 module.exports = {
   getBreadcrumbs,
-  getFlatSidebar,
 };
