+++ src/utils/get-flat-sidebar.js
@@ -1,0 +1,12 @@
+const getFlatSidebar = (sidebar, path = []) =>
+  sidebar.reduce((acc, item, index) => {
+    const current = { title: item.title, slug: item.slug, path: [...path, index] };
+    if (item.items) {
+      return [...acc, current, ...getFlatSidebar(item.items, current.path)];
+    }
+    return [...acc, { ...item, path: [...path, index] }];
+  }, []);
+
+module.exports = {
+  getFlatSidebar,
+};
